==> code/1_preprocessing/3_preprocessing_helm_dexi.ipynb <==
# 3. Preprocessing JAEB T1-DEXI dataset

The goal of this notebook is to prepare the EXTOD education data for machine learning - predicting euglycemia during and around exercise.

EXTOD education was a pilot study in which 106 participants were randomly allocated to either standard care or an education programme teaching them how better to manage blood glucose around exercise.

The data used will be demographic, lab, clinical and physiological if available.

### Objectives:
1. Clean and combine the exercise diaries
2. Preprocess demographic data
3. Preprcoess lab data
4. Prepare CGM data

## 3.0. Import packages

# Import packages and upload dataset
import pandas as pd
from datetime import timedelta as time
import warnings
import preprocess_helper
import functools as ft

warnings.filterwarnings('ignore')

# Set the directory
directory = '../../data/raw_data/helmsley/'

## 3.1. Clean and combine exercise diaries

### 3.1.1. Load files and rename columns

# Read file
exercise = pd.read_csv(directory+'pr_dexi.csv')

# Convert the start time to datetime
exercise['PRSTDTC'] = pd.to_datetime(exercise['PRSTDTC'])

exercise

# Select and rename columns
exdi = exercise[['USUBJID', 'PRSTDTC', 'PLNEXDUR', 'EXCINTSY', 'PRTRTC']] 
exdi.columns = ['ID', 'start_datetime', 'duration', 'intensity', 'type_of_exercise']

exdi.dropna(subset=['ID'], inplace=True)

# ID as int for concatenation
exdi['ID'] = exdi['ID'].astype(int)

exdi.shape

len(exdi.ID.unique())

### 3.1.2. Adjust date columns

# Calculate end time from start and duration 
exdi['finish_datetime'] = exdi.apply(lambda row: row['start_datetime'] + time(minutes=row['duration']), axis=1)

# Collect date information from start_datetime
exdi = preprocess_helper.date_preprocessing(exdi, 'start_datetime', False, True, True)

# Drop any instances with no id, start time or duration
exdi = exdi.dropna(subset=['ID', 'start_datetime', 'duration'])

# Make sure there's no minus durations
exdi = exdi.loc[exdi['duration']>0]

exdi.shape[0]

def handle_overlaps(group):
    # Sort by start_datetime
    group = group.sort_values(by='start_datetime')
    
    non_overlapping = []
    overlap_count = 0  # Counter to limit the number of printed overlaps
    for _, row in group.iterrows():
        if non_overlapping:
            last_interval = non_overlapping[-1]
            # Check for overlap
            if row['start_datetime'] < last_interval['finish_datetime']:
                # Print the overlapping intervals (limited to a few for brevity)
                if overlap_count < 5:
                    print(f"Overlap detected for ID {group['ID'].iloc[0]}:")
                    print(f"Interval 1: {last_interval['start_datetime']} to {last_interval['finish_datetime']}")
                    print(f"Interval 2: {row['start_datetime']} to {row['finish_datetime']}\n")
                    overlap_count += 1

                curr_duration = (row['finish_datetime'] - row['start_datetime']).seconds / 60
                last_duration = (last_interval['finish_datetime'] - last_interval['start_datetime']).seconds / 60
                
                # Decide which interval to keep
                if (curr_duration >= 10 and curr_duration <= 120) and (last_duration < 10 or last_duration > 120):
                    non_overlapping[-1] = row
                elif (last_duration >= 10 and last_duration <= 120) and (curr_duration < 10 or curr_duration > 120):
                    continue
                elif curr_duration > last_duration:
                    non_overlapping[-1] = row
            else:
                non_overlapping.append(row)
        else:
            non_overlapping.append(row)
    
    return pd.DataFrame(non_overlapping)

# Apply the function to each group
exdi = exdi.groupby('ID').apply(handle_overlaps).reset_index(drop=True)


exdi.shape[0]

### 3.1.3. Type of exercise

# Convert form_of_exercise
exdi['form_of_exercise'] = exdi.type_of_exercise.apply(lambda x: 
                                                       preprocess_helper.divide_exercise_into_type(x))

# Removing bouts that aren't really exercise
#exdi.dropna(subset=['form_of_exercise'], inplace=True)

exdi.shape

## 3.2. Clean and combine CGM files

# Upload directory for CGM data
lb = pd.read_csv(directory+'lb_dexi.csv')

# Select CGM readings
cgm = lb.loc[lb['LBCAT']=='CGM']

# Convert timestamp to datetime
cgm['LBDTC'] = pd.to_datetime(cgm['LBDTC'])

# Select and rename columns 
cgm = cgm[['USUBJID', 'LBDTC', 'LBORRES']]
cgm.columns = ['ID', 'time', 'glc']

# Drop any null 
cgm.dropna(inplace=True)

# Convert glucose to mg/dl
cgm['glc'] = (cgm['glc']/18).round(2)

# ID to int for concatenation
cgm['ID'] = cgm['ID'].astype(int)


## 3.3. Lab & demographic data

# Read file
demographics = pd.read_csv(directory+'dm_dexi.csv')

# Select and rename columns
demo = demographics[['USUBJID', 'AGE', 'SEX', 'RACE']]
demo.columns = ['ID', 'age', 'sex', 'race']

# Replace m/f values
demo['sex'] = demo['sex'].replace({'M':'male', 'F':'female'})

# ID as int for concatenation
demo['ID'] = demo['ID'].astype(int)

### 3.3.1. Calculate hba1c

# Select the hba1c results from the tests file
hba1c = lb.loc[lb['LBTESTCD']=='HBA1C']

# Make id int for concatenation
hba1c['USUBJID'] = hba1c['USUBJID'].astype(int)


# Select and rename columns
hba1c = hba1c[['USUBJID', 'LBORRES']]
hba1c.columns = ['ID', 'hba1c']

# Convert to mmol/mol
hba1c['hba1c'] =(hba1c['hba1c'] -2.15)*10.929



### 3.3.2. Calculate diabetes onset

# Set up list for results
fa_data = []

# Read file
fa = pd.read_sas(directory+'FA.xpt', encoding='utf-8',chunksize=10000, iterator=True)

# Read the file in chunks because it's so massive
for chunk in fa:
    chunk = chunk.loc[chunk['FAOBJ']== "DIABETES ONSET"]
    if chunk.shape[0]==0:
        continue
    fa_data.append(chunk[['USUBJID', 'FASTRESN']])

# Concat all the chunks
fa_data=pd.concat(fa_data)

# Rename coluns
fa_data.columns = ['ID', 'years_since_diagnosis']

# Make id int for concatenation
fa_data['ID'] = fa_data['ID'].astype(int)

### 3.3.3. Calculate BMI

# Set up list for results
vs_data=[]

# Read file
vs = pd.read_sas(directory+'VS.xpt', encoding='utf-8',chunksize=10000, iterator=True)

# Read the file in chunks because it's so massive
for chunk in vs:
    chunk = chunk.loc[(chunk['VSTESTCD']== 'HEIGHT') | (chunk['VSTESTCD']=='WEIGHT')]
    if chunk.shape[0]==0:
        continue
    vs_data.append(chunk[['USUBJID', 'VSTESTCD', 'VSORRES']])

# Concat all the chunks
vs_data= pd.concat(vs_data)

# Make it the right way round
vs_data_adj = vs_data.pivot(columns='VSTESTCD', index='USUBJID')

vs_data_adj.columns = ['h', 'w']

#Calculate bmi
vs_data_adj['bmi'] = (703*vs_data_adj['w'])/(vs_data_adj['h']*vs_data_adj['h'])

# Reset index
vs_data_adj = vs_data_adj.reset_index()

# Select only id and bmi and rename
bmi = vs_data_adj[['USUBJID', 'bmi']]
bmi.columns = ['ID', 'bmi']

# Make id int for concatenatin
bmi['ID'] = bmi['ID'].astype(int)

## 3.4. Insulin modality

# Read file
dx = pd.read_sas(directory+'DX.xpt', encoding='utf-8')

# Rename values
insulin_modality = dx[dx['DXCAT']=='INSULIN MODALITY'][['USUBJID', 'DXTRT']].replace({'MULTIPLE DAILY INJECTIONS': 'mdi', 
                                                                                      'INSULIN PUMP':'pump', 
                                                                                      'CLOSED LOOP INSULIN PUMP':'closed_loop'})

# Rename columns
insulin_modality.columns = ['ID', 'insulin_modality']

# Convert ID to int for concatenation
insulin_modality['ID'] = insulin_modality['ID'].astype(int)

## 3.5. Combine all dfs

# List of all dfs
dfs = [demo, hba1c, bmi, fa_data, insulin_modality]

# Merge them all
demo_lab = ft.reduce(lambda left, right: pd.merge(left, right, on='ID', how='outer'), dfs)

# Add 'helm' to IDs to make them unique
for i in [exdi, cgm, demo_lab]:
    i['ID'] = i['ID'].apply(lambda row: 'helm_'+str(row))

## 3.6. Save files

df_directory = '../../data/tidy_data/helmsley/'
exdi.to_csv(df_directory+'exercise.csv', index=False)
cgm.to_csv(df_directory+'cgm.csv', index=False)
demo_lab.to_csv(df_directory+'demo_lab.csv', index=False)
