==> code/1_preprocessing/6_ml_preparation.ipynb <==
# 6. Final preparation for machine learning
Before delving into the machine learning modeling, it's pivotal to ensure that the data is adequately cleaned, processed, and formatted. This section delineates the steps undertaken to finalize the data in preparation for model training and validation.
Highlights:

* Data Importing: Leveraging libraries such as pandas and numpy, the DEXIP study data is loaded and initialized.

* Target Variables: Two distinct binary target variables (y & y_3) are crafted, stemming from glucose level thresholds. Additionally, a 'stratify' column is introduced, marrying participant ID with the y_3 target, to aid in stratified sampling later on.

* Data Trimming: Emphasis is placed on data quality. Rows deficient in data (less than 60% sufficiency) are pruned, and columns with over 20% missing data are cast aside. Furthermore, rows lacking key columns like start_glc and duration are eradicated.

* Data Refinement: Innovative features like 'season' are derived from existing ones, offering richer data for modeling. Certain columns, mostly timestamps, are processed and converted to a numeric format.

* Tidying & Saving Data: A series of operations—including resetting indices, discarding superfluous columns, replacing infinite values, creating dummy variables for categorical columns, and employing KNN imputation for missing values—culminates in the final data. The finished dataframe, along with other critical components like feature matrix (X) and target vector (y), is archived for subsequent use.

With the data now in its prime, it paves the way for the ensuing phase: machine learning modeling.

## 6.0. Import packages and load data

import pandas as pd
import numpy as np
import preprocess_helper
import sys
from sklearn.impute import KNNImputer
path = "../../diametrics"
sys.path.append(path)

df = pd.read_csv('../../data/tidy_data/data_with_targets_dexip.csv')

df[(df['duration']>=10) | (df['duration']<=120)].shape

[i for i in df.columns]

## 6.1. Prepare target variables

# Target variable for 5mmol/L
df['y'] = df['tir_below_5']>0

# Target variable for 3.9mmol/L
df['y_3'] = df['tir_below_3']>0

# Create column for stratifying based on ID and target (3.9)
df['stratify'] = df['ID']+ '_'+ df['y'].astype(str)

## 6.2. Cutting data

# Drop any instances without start glucose and duration
df = df.dropna(subset=['start_glc', 'duration']).reset_index()

df.shape

# Drop columns with < 60% data sufficiency
df = df.loc[df['data_suff_during']>=60]

# Drop columns with < 60% data sufficiency
df = df.loc[df['before_data_suff']>=40]

df.shape

# Drop columns with less than 80% data available
df_dropped = df.loc[:, df.isnull().mean() < .2]

df.shape

df.shape

## 6.3. Data adjustments

# Create 'season' variable
df['season'] = (df.month%12 // 3 + 1).astype(str)

# Make datetime cols numeric
cols = ['Avg. length of hypoglycemic events', 
        'Total time spent in hypoglycemic events',
         'Avg. length of hyperglycemic events',
         'Total time spent in hyperglycemic events', ]

for col in cols:
    df_dropped[col] = df_dropped[col].apply(lambda x: preprocess_helper.make_numeric(x)).astype(float)

## Get rid of weird bouts

# Convert form_of_exercise
df_dropped['form_of_exercise'] = df_dropped.type_of_exercise.apply(lambda x: 
                                                       preprocess_helper.divide_exercise_into_type(x))

df_dropped = df_dropped.dropna(subset=['form_of_exercise'])

df_dropped.shape


df_dropped=df_dropped.drop_duplicates(['bout_id', 'start_datetime'])

df_dropped.shape

## 6.4. Adjust and tidy data

# Save df for demographics table
df_dropped.to_csv('../../data/tidy_data/demographics_df.csv', index=False)

# Save id, bout id etc as a key
df_dropped[['ID', 'bout_id', 'start_datetime','finish_datetime']].to_csv('../1_preprocessing/id_key_dexip.csv', index=False)

# Reset index after dropping
df_dropped.reset_index(inplace=True, drop=True)

# Drop columns that arent needed
cols_to_drop = ['start_datetime', 'finish_datetime', 'type_of_exercise',
               'data_suff_during', # 'interval', 'tir_below_5', 'tir_below_3',
             'month', 'day',  'ID',  'index']

df = df_dropped.drop(columns=cols_to_drop)

# Repace infs in tsfresh with nan
df.replace([np.inf, -np.inf], np.nan, inplace=True)

# Create X 
X = df.drop(columns=['study','interval', 'stratify', 'bout_id', 'y'])

# Get dummies
X = pd.get_dummies(X)

# Fill missing data
imputer = KNNImputer(n_neighbors=5, missing_values=np.nan)
X = pd.DataFrame(imputer.fit_transform(X), columns=X.columns)

# Round intensity back to ordinal after imputing
X.intensity = X.intensity.round()

# Set y as 3.9mmol/L
y = df['y']

# Save everything
df.to_csv('../../data/tidy_data/final_df/df.csv', index=False)
df['stratify'].to_csv('../../data/tidy_data/final_df/strat.csv', index=False)
X.to_csv('../../data/tidy_data/final_df/X.csv', index=False)
y.to_csv('../../data/tidy_data/final_df/y.csv', index=False)
