==> code/1_preprocessing/7_participant_characteristics.ipynb <==
# 7. Analysis of Participant Characteristics in Exercise Studies

This Jupyter Notebook provides a comprehensive analysis of participant characteristics across various exercise studies. The analysis encompasses both demographic and exercise bout data, as well as metrics of glycemic control derived from continuous glucose monitoring (CGM) data.

### Structure of the Notebook:

1. Import Packages and Upload Data
        The necessary Python packages are imported for data processing, statistical analyses, and visualization.
        Raw data files are uploaded from specified directories, and an initial dataset df is created, which is further enhanced with the addition of id_key and race data from DEXI and DEXIP studies.
        Appropriate data directories for different studies are specified.

2. Data Dictionary
        The data is described in terms of its structure and content.
        Data types, missing values, and basic statistics (min, max, quartiles) for numeric columns are calculated.
        Unique values and categories for categorical columns are identified.

3. Statistical Analysis of Demographic and Exercise Bout Data
        Lists of categorical and numeric variables related to exercise bouts and demographics are defined.
        Summary statistics, including medians and interquartile ranges, are computed for each study as well as for all studies combined.

4. Metrics of Glycemic Control
        Glycemic metrics for each study (T1-DEXI, T1-DEXIP, EXT-101, EXT-edu) are calculated using the metrics.all_standard_metrics function.
        Results from individual studies are combined into a single dataset all_metrics.
        Summary statistics for these metrics are again computed for each study and for the combined dataset.

5. Compilation and Presentation of Results
        The results from the demographic/exercise bout data analysis and the glycemic control metrics analysis are combined.
        A final table, ci_table, is created, which provides a comparative overview of the results across all studies.
        This table is then saved to a CSV file for external use.

### Data Sources:

Demographics Data: demographics_df.csv
ID Key for DEXIP Study: id_key_dexip.csv
Race Data for DEXI and DEXIP Studies: dexi_race.csv, dexip_race.csv
CGM Data for Various Studies: Located in specified directories for each study.

### Objective:

The goal of this analysis is to provide a clear understanding of the participant characteristics across the different studies, focusing on demographics, exercise bout information, and glycemic control metrics. This consolidated view aids in drawing meaningful insights and comparisons between studies and helps identify patterns or trends related to exercise and glycemia.

## 7.0 Import packages and upload data

import pandas as pd
import numpy as np
from datetime import datetime, date, timedelta
from functools import reduce
import preprocess_helper
import sys
path = "../../diametrics"
sys.path.append(path)
import metrics

import warnings
warnings.filterwarnings('ignore')


# Directories
directory_101 = '../../data/tidy_data/extod_101/'
directory_edu = '../../data/tidy_data/extod_edu/'
directory_helm = '../../data/tidy_data/helmsley/'
directory_dexip = '../../data/tidy_data/dexip/'

# Upload df 
df = pd.read_csv('../../data/tidy_data/demographics_df.csv')

df.shape

# Upload ID key and merge with df
id_key = pd.read_csv('../1_preprocessing/id_key_dexip.csv').drop_duplicates()
df = df.merge(id_key, on='bout_id', how='inner')

id_key

## 7.1. Data dictionary

# Convert ordinal variables to str
cols_to_convert = ['intensity', 'day_of_week'] #'season', 
for col in cols_to_convert:
    df[col] = df[col].astype(str)

# Calculate relevant info for data dictionary
numeric_dict = {}
categorical_dict = {}

for column in df.columns:
    missing_values = df[column].isnull().sum()

    if df[column].dtype in ['int64', 'float64']:  # Numeric columns
        numeric_dict[column] = {
            'data_type': str(df[column].dtype),
            'missing_values': missing_values,
            'min': df[column].min(),
            'Q1': df[column].quantile(0.25),
            'Q2': df[column].median(),
            'Q3': df[column].quantile(0.75),
            'max': df[column].max()
        }
    else:  # Categorical columns
        categorical_dict[column] = {
            'data_type': str(df[column].dtype),
            'unique_values': df[column].nunique(),
            'missing_values': missing_values,
            'categories': df[column].unique()#.tolist()
        }


# Save categorical data
cat_data = pd.DataFrame(categorical_dict).T
cat_data.to_csv('../../results/data_dictionary/categorical_data.csv')

cat_data

num_data = pd.DataFrame(numeric_dict).T

num_data.to_csv('../../results/data_dictionary/numeric_data.csv')

## 7.2. Statistical analysis

### 7.2.1. Demographic and exercise bout data

# Categorical variables relating to exercise bout
cat_bout = ['intensity',
               'day_of_week',
               'time_of_day',
               'form_of_exercise',
               ]

# Numeric variables relating to exercise bout
numeric_bout = ['duration',
    'start_glc',
    'start_roc',]

# Categorical variables relating to demographics
cat_demo = [
    'sex',
    'insulin_modality',
    'race'
]

# Numeric variables relating to demographics
numeric_demo = [
    'age',
    'hba1c',
    'bmi',
    'years_since_diagnosis',
    ]

# Calculate relevant info for each study
overview = df.groupby('study').apply(lambda group: preprocess_helper.overview_results(group,  numeric_demo, cat_demo, numeric_bout, cat_bout))
overview = overview.reset_index().drop(columns='level_1')

# Calculate relevant info for all studies combined
all_studies = preprocess_helper.overview_results(df,  numeric_demo, cat_demo, numeric_bout, cat_bout)
all_studies['study'] = 'all'
all_results = all_studies.append(overview)

# Rename cols and reset index
all_results.columns = ['Feature', 'Result (Median [IQR])', 'Study']
all_results = all_results.reset_index(drop=True)

### 7.2.2 Metrics of glycemic control 

# T1-DEXI metrics
helmsley_cgm = pd.read_csv(directory_helm + 'cgm.csv')
helmsley_cgm['time'] = pd.to_datetime(helmsley_cgm['time'])
helm_metrics = metrics.all_standard_metrics(helmsley_cgm)

# T1-DEXIP metrics
dexip_cgm = pd.read_csv(directory_dexip + 'cgm.csv')
dexip_cgm['time'] = pd.to_datetime(dexip_cgm['time'])
dexip_metrics = metrics.all_standard_metrics(dexip_cgm)

# EXT-101 metrics
extod_101_cgm = pd.read_csv(directory_101 + 'cgm.csv')
extod_101_cgm['time'] = pd.to_datetime(extod_101_cgm['time'])
ext_101_metrics = metrics.all_standard_metrics(extod_101_cgm)

# EXT-edu metrics
extod_edu_cgm = pd.read_csv(directory_edu + 'cgm.csv')
extod_edu_cgm['time'] = pd.to_datetime(extod_edu_cgm['time'])
ext_edu_metrics = metrics.all_standard_metrics(extod_edu_cgm)

# Combine all studies
all_metrics = helm_metrics.append(dexip_metrics).append(ext_edu_metrics).append(ext_101_metrics)

# Save the metrics
all_metrics.to_csv('../../results/metrics_glycemic_control.csv')

all_metrics = pd.read_csv('../../results/metrics_glycemic_control.csv')

all_metrics

# Drop columns that won't be included in table
cols = ['Avg. length of hypoglycemic events', 
        'Total time spent in hypoglycemic events',
         'Avg. length of hyperglycemic events',
         'Total time spent in hyperglycemic events','ID',
         'Unnamed: 0' ]
all_metrics_dropped  = all_metrics.drop(columns=cols)
remaining_cols = all_metrics_dropped.columns

len(all_metrics.ID.unique())

# Calculate mean[iqr] glycemic metrics for all participants 
all_metrics_iqr = preprocess_helper.col_iqr(all_metrics, remaining_cols)
all_metrics_iqr['Study'] = 'all'

# Calculate mean[iqr] glycemic metrics for dexi participants 
helm_metrics_iqr =  preprocess_helper.col_iqr(all_metrics[all_metrics['ID'].str.startswith('helm')], remaining_cols)
helm_metrics_iqr['Study'] = 'helm'

# Calculate mean[iqr] glycemic metrics for dexip participants 
dexip_metrics_iqr =  preprocess_helper.col_iqr(all_metrics[all_metrics['ID'].str.startswith('dexip')], remaining_cols)
dexip_metrics_iqr['Study'] = 'dexip'

# Calculate mean[iqr] glycemic metrics for edu participants 
ext_edu_metrics_iqr =  preprocess_helper.col_iqr(all_metrics[all_metrics['ID'].str.startswith('ext_edu')], remaining_cols)
ext_edu_metrics_iqr['Study'] = 'ext_edu'

# Calculate mean[iqr] glycemic metrics for 101 participants 
ext_101_metrics_iqr =  preprocess_helper.col_iqr(all_metrics[all_metrics['ID'].str.startswith('ext_101')], remaining_cols)
ext_101_metrics_iqr['Study'] = 'ext_101'

# Combine all the studies cgm data
overview_cgm = all_metrics_iqr.append(helm_metrics_iqr).append(dexip_metrics_iqr).append(ext_edu_metrics_iqr).append(ext_101_metrics_iqr)

# Rename cols
overview_cgm.columns = ['Feature', 'Result (Median [IQR])', 'Study']


### 7.2.3. Combine datasets

# Add CGM results to demographic and exercise data
all_results = all_results.append(overview_cgm)

# Display table
ci_table = all_results.pivot(columns='Feature', values='Result (Median [IQR])', index='Study').T

ci_table_ordered = ci_table[['all','helm',	'dexip',	'ext_101',	'ext_edu']].reset_index()

# Conversion function
def convert_to_mg_per_dl(value_str):
    # Conversion factor from mmol/L to mg/dL
    conversion_factor = 18.01559
    
    # Extract the main value and the range values
    main_value = float(value_str.split('[')[0].strip())
    range_values = [float(val) for val in value_str.split('[')[1].replace(']', '').split(',')]
    
    # Convert the values
    main_value_mg_per_dl = round(main_value * conversion_factor, 1)
    range_values_mg_per_dl = [round(val * conversion_factor, 1) for val in range_values]
    
    # Format the result
    result = "{} [{} , {}]".format(main_value_mg_per_dl, range_values_mg_per_dl[0], range_values_mg_per_dl[1])
    
    return result

# List of features to convert
features_to_convert = ["Average glucose (mmol/L)", "SD (mmol/L)", "start_glc"]

# Columns to apply the conversion to
cols_to_convert = ['all', 'helm', 'dexip', 'ext_101',	'ext_edu']  # Add other columns as needed

# Apply the function to the specified rows and columns
for feature in features_to_convert:
    ci_table_ordered.loc[ci_table_ordered['Feature'] == feature, cols_to_convert] = ci_table_ordered.loc[ci_table_ordered['Feature'] == feature, cols_to_convert].applymap(convert_to_mg_per_dl)

print(ci_table_ordered)

ci_table_ordered

# Save table
ci_table.to_csv('../../results/demographics_table.csv')
