==> code/1_preprocessing/8_consort_diagram.ipynb <==
import rpy2.robjects as robjects

# R's utility function
utils = robjects.packages.importr('utils')

# Select a CRAN mirror
utils.chooseCRANmirror(ind=1)

# Install the package
utils.install_packages('consortDiagram')

import rpy2.robjects as robjects
import rpy2.robjects.packages as rpackages
# Install the package
rpackages.install_packages('consortDiagram')
# Initialize R function
robjects.r('''
    install.packages("consortDiagram")
    library(consortDiagram)
    create_consort <- function() {
        # Your R code to create the consort diagram using consortDiagram package
        # For example:
        cd <- consortDiagram$new()
        cd$add("Total", 1000)
        cd$drop("Excluded", 100)
        cd$render()
    }
''')

# Call the R function from Python
create_consort = robjects.globalenv['create_consort']
create_consort()
